third_party/blink/web_tests/accessibility/input-type-text-caret-position-expected.txt

This tests that caret position is reported correctly for text input elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS inputAccessible is inputAccessible.selectionAnchorObject
PASS inputAccessible.selectionAnchorOffset is 0
PASS inputAccessible is inputAccessible.selectionFocusObject
PASS inputAccessible.selectionFocusOffset is 0
PASS inputAccessible is inputAccessible.selectionAnchorObject
PASS inputAccessible.selectionAnchorOffset is 1
PASS inputAccessible is inputAccessible.selectionFocusObject
PASS inputAccessible.selectionFocusOffset is 1
PASS inputAccessible is inputAccessible.selectionAnchorObject
PASS inputAccessible.selectionAnchorOffset is 2
PASS inputAccessible is inputAccessible.selectionFocusObject
PASS inputAccessible.selectionFocusOffset is 2
PASS inputAccessible is inputAccessible.selectionAnchorObject
PASS inputAccessible.selectionAnchorOffset is 3
PASS inputAccessible is inputAccessible.selectionFocusObject
PASS inputAccessible.selectionFocusOffset is 3
PASS inputAccessible is inputAccessible.selectionAnchorObject
PASS inputAccessible.selectionAnchorOffset is 4
PASS inputAccessible is inputAccessible.selectionFocusObject
PASS inputAccessible.selectionFocusOffset is 4
PASS inputAccessible is inputAccessible.selectionAnchorObject
PASS inputAccessible.selectionAnchorOffset is 5
PASS inputAccessible is inputAccessible.selectionFocusObject
PASS inputAccessible.selectionFocusOffset is 5
PASS inputAccessible is inputAccessible.selectionAnchorObject
PASS inputAccessible.selectionAnchorOffset is 6
PASS inputAccessible is inputAccessible.selectionFocusObject
PASS inputAccessible.selectionFocusOffset is 6
PASS successfullyParsed is true

TEST COMPLETE
